+++ CHANGES.rst
@@ -11,6 +11,10 @@ New Features in 25.1
 
 - Guermok HDMI to USB 3.0 capture dongle supported
 
+- The `USBVideoDriver`'s ``stream`` method can now be called with ``fps=True``
+  to embed an overlay with FPS statistics into the video stream.
+  This is exposed in ``labgrid-client video`` via the new ``--fps`` option.
+
 Breaking changes in 25.1
 ~~~~~~~~~~~~~~~~~~~~~~~~
 - The deprecated pytest plugin option ``--env-config`` has been removed. Use